+++ service.sh
@@ -1,7 +1,8 @@
 MODPATH=${0%/*}
 
 # log
-exec 2>$MODPATH/debug.log
+LOGFILE=$MODPATH/debug.log
+exec 2>$LOGFILE
 set -x
 
 # var
@@ -57,12 +58,9 @@ resetprop vendor.audio.use.sw.alac.decoder true
 # special file
 FILE=/dev/sony_hweffect_params
 FILE2=/dev/msm_hweffects
-FILE3=/dev/mtk_snd_soc_sounddev
 if [ ! -e $FILE ]; then
   if [ -e $FILE2 ]; then
     MM=`stat -c "%t %T" $FILE2 | { read major minor; printf "%d %d\n" 0x$major 0x$minor; }`
-  elif [ -e $FILE3 ]; then
-    MM=`stat -c "%t %T" $FILE3 | { read major minor; printf "%d %d\n" 0x$major 0x$minor; }`
   fi
   if [ "$MM" ]; then
     mknod $FILE c $MM
@@ -80,7 +78,7 @@ else
 fi
 PID=`pidof $SERVER`
 if [ "$PID" ]; then
-  killall $SERVER
+  killall $SERVER android.hardware.audio@4.0-service-mediatek
 fi
 
 # unused
@@ -129,7 +127,8 @@ killall vendor.qti.hardware.vibrator.service\
  android.hardware.sensors@1.0-service\
  android.hardware.sensors@2.0-service\
  android.hardware.sensors@2.0-service-mediatek\
- android.hardware.sensors@2.0-service.multihal
+ android.hardware.sensors@2.0-service.multihal\
+ android.hardware.health-service.qti
 }
 
 # dolby
@@ -204,7 +203,7 @@ fi
 
 # allow
 PKG=com.dolby.daxappui
-if pm list packages | grep $PKG; then
+if appops get $PKG > /dev/null 2>&1; then
   if [ "$API" -ge 30 ]; then
     appops set $PKG AUTO_REVOKE_PERMISSIONS_IF_UNUSED ignore
   fi
@@ -217,7 +216,7 @@ fi
 
 # allow
 PKG=com.dolby.daxservice
-if pm list packages | grep $PKG; then
+if appops get $PKG > /dev/null 2>&1; then
   if [ "$API" -ge 30 ]; then
     appops set $PKG AUTO_REVOKE_PERMISSIONS_IF_UNUSED ignore
   fi
@@ -230,10 +229,10 @@ fi
 
 # function
 stop_log() {
-FILE=$MODPATH/debug.log
-SIZE=`du $FILE | sed "s|$FILE||g"`
+SIZE=`du $LOGFILE | sed "s|$LOGFILE||g"`
 if [ "$LOG" != stopped ] && [ "$SIZE" -gt 50 ]; then
   exec 2>/dev/null
+  set +x
   LOG=stopped
 fi
 }
@@ -247,15 +246,11 @@ sleep 15
 stop_log
 NEXTPID=`pidof $SERVER`
 if [ "`getprop init.svc.$SERVER`" != stopped ]; then
-  until [ "$PID" != "$NEXTPID" ]; do
-    check_audioserver
-  done
-  killall $PROC
-  check_audioserver
+  [ "$PID" != "$NEXTPID" ] && killall $PROC
 else
   start $SERVER
-  check_audioserver
 fi
+check_audioserver
 }
 check_service() {
 for SERVICE in $SERVICES; do
